+++ mp3voicestamp-app/app.py
@@ -17,11 +17,11 @@ from __future__ import print_function
 
 import os
 import shutil
-import signal
 import tempfile
 from collections import OrderedDict
 
 from args import Args
+from job import Job
 from util import Util
 
 
@@ -38,14 +38,6 @@ class App(object):
 
     def __init__(self):
         self.args = None
-        self.tmp_dir = None
-
-    def sigint_handler(self, unused_signal, unused_frame):
-        self.cleanup()
-
-    def cleanup(self):
-        if self.args.no_cleanup and self.tmp_dir is not None:
-            shutil.rmtree(self.tmp_dir)
 
     def get_mp3_info(self, mp3_file_name):
         from mutagen.mp3 import MP3
@@ -85,46 +77,38 @@ class App(object):
         tools = ['ffmpeg', 'normalize-audio', 'espeak', 'sox']
         for tool in tools:
             if Util.which(tool) is None:
-                Util.abort('Missing "{}". See README.md for list of mandatory dependencies.'.format(tool))
+                Util.abort('"{}" not found. See README.md for details.'.format(tool))
 
     def speak_to_wav(self, text, out_file_name):
         rc = Util.execute_rc(['espeak', '-s', '150', '-z', '-w', out_file_name, str(text)])
         return rc == 0
 
-    def main(self):
+    def process_file(self, job):
         import wave
 
-        self.check_env()
-
-        # parse comman line arguments
-        self.args = Args.parse_args()
-        Util.init(self.args)
-
         # ensure it's safe to write use give name for output file
-        if os.path.exists(self.args.file_out):
-            if not self.args.force:
-                Util.abort('Target "{}" already exists'.format(self.args.file_out))
-            if not os.path.isfile(self.args.file_out):
-                Util.abort('Target "{}" is not a file. Cannot overwrite'.format(self.args.file_out))
+        if os.path.exists(job.file_out):
+            if not job.force:
+                Util.abort('Target "{}" already exists'.format(job.file_out))
+            if not os.path.isfile(job.file_out):
+                Util.abort('Target "{}" is not a file. Cannot overwrite'.format(job.file_out))
 
         # get some info about source MP3 file and print it
-        in_file_info = self.get_mp3_info(self.args.file_in)
-        Util.print_info(in_file_info, 'Input track info')
+        in_file_info = self.get_mp3_info(job.file_in)
+        if job.args.verbose:
+            Util.print_info(in_file_info, 'Input track info')
 
         # some sanity checks.
-        min_track_length = 1 + self.args.tick_offset
+        min_track_length = 1 + job.tick_offset
         if in_file_info['duration'] < min_track_length:
             Util.abort('Input MP3 track must be at least {} minutes long'.format(min_track_length))
 
         # create temp folder where all out working files will live for a while
-        self.tmp_dir = tempfile.mkdtemp()
-        Util.debug('Temporary folder: {}'.format(self.tmp_dir))
-
-        # plant SIGINT handler, so we can clean up temp folder when interrupted
-        signal.signal(signal.SIGINT, self.sigint_handler)
+        tmp_dir = tempfile.mkdtemp()
+        Util.debug('Temporary folder: {}'.format(tmp_dir))
 
         # prepare some placeholders to be later use for speaking track title.
-        track_title = self.args.title_pattern.format(
+        track_title = job.title_pattern.format(
             title=in_file_info['title'],
             artist=in_file_info['artist'],
             album_artist=in_file_info['album_artist'],
@@ -135,22 +119,23 @@ class App(object):
         )
 
         # let's see how many time stamps we are going to have
-        ticks = range(self.args.tick_offset,
+        ticks = range(job.tick_offset,
                       in_file_info['duration'],
-                      self.args.tick_interval)
+                      job.tick_interval)
         out_file_info = OrderedDict([
-            ('file_name', self.args.file_out),
+            ('file_name', job.file_out),
             ('quality', self.guess_encoding_quality_for_lame_encoder(in_file_info['bitrate'])),
-            ('title_pattern', self.args.title_pattern),
+            ('title_pattern', job.title_pattern),
             ('title_overlay', track_title),
-            ('tick_pattern', self.args.tick_pattern),
-            ('tick_offset', self.args.tick_offset),
-            ('tick_interval', self.args.tick_interval),
-            ('tick_volume_factor', self.args.tick_volume_factor),
+            ('tick_pattern', job.tick_pattern),
+            ('tick_offset', job.tick_offset),
+            ('tick_interval', job.tick_interval),
+            ('tick_volume_factor', job.tick_volume_factor),
             ('tick_count', len(ticks))
         ])
         # print some info about what we are going to produce and how
-        Util.print_info(out_file_info, 'Output track info')
+        if job.args.verbose:
+            Util.print_info(out_file_info, 'Output track info')
 
         # let's now create WAVs with our spoken parts. First goes track title, then time stamps
         segments = [track_title]
@@ -159,15 +144,15 @@ class App(object):
             segments.append(out_file_info['tick_pattern'].format(time_marker))
 
         for idx, segment_text in enumerate(segments):
-            segment_file_name = os.path.join(self.tmp_dir, '{}.wav'.format(idx))
+            segment_file_name = os.path.join(tmp_dir, '{}.wav'.format(idx))
             if not self.speak_to_wav(segment_text, segment_file_name):
                 Util.abort('Failed to create "{0}" as "{1}".'.format(segment_text, segment_file_name))
         Util.print('OK')
 
         # we need to get the frequency of speech waveform generated by espeak to later be able to tell
         # ffmpeg how to pad/clip the part
-        wav = wave.open(os.path.join(self.tmp_dir, '0.wav'), 'rb')
-        speech_meged_wav_full = os.path.join(self.tmp_dir, 'speech.wav')
+        wav = wave.open(os.path.join(tmp_dir, '0.wav'), 'rb')
+        speech_meged_wav_full = os.path.join(tmp_dir, 'speech.wav')
         speech_frame_rate = wav.getframerate()
         wav.close()
 
@@ -181,7 +166,7 @@ class App(object):
         max_len_tick = speech_frame_rate * 60 * out_file_info['tick_interval']
         max_len_title = speech_frame_rate * 60 * out_file_info['tick_offset']
         for idx, i in enumerate(segments):
-            concat_cmd.extend(['-i', os.path.join(self.tmp_dir, '{}.wav'.format(idx))])
+            concat_cmd.extend(['-i', os.path.join(tmp_dir, '{}.wav'.format(idx))])
 
             # samples = rate_per_second * seconds * tick_interval_in_minutes
             max_len = max_len_title if idx == 0 else max_len_tick
@@ -204,15 +189,15 @@ class App(object):
         # this is needed for many reasons:
         # we need to adjust voice overlay amplitude to match MP3 file level and to do that we use "sox" too
         # which cannot deal with MP3 directly.
-        # Also there are some odd issues with "ffmpeg" failing during merge when source was in mp3 - blid guess for now
-        # it was due to some struct mismatch between MP3 file (i.e. having cover image), and speech segments being
+        # Also there are some odd issues with "ffmpeg" failing during merge when source was in mp3 - blind guess for now
+        # it was due to some structure mismatch between MP3 file (i.e. having cover image), and speech segments being
         # just waveform. Most likely this can be solved better way but we need WAV anyway so no point wasting time
         # at the moment.
 
-        source_wav_file_name = self.args.file_in + '.wav'
-        source_wav_full_path = os.path.join(self.tmp_dir, source_wav_file_name)
+        source_wav_file_name = os.path.basename(job.file_in) + '.wav'
+        source_wav_full_path = os.path.join(tmp_dir, source_wav_file_name)
         Util.print_no_lf('Creating temporary WAV')
-        wav_cmd = ['ffmpeg', '-i', self.args.file_in, source_wav_full_path]
+        wav_cmd = ['ffmpeg', '-i', job.file_in, source_wav_full_path]
         if Util.execute_rc(wav_cmd) != 0:
             Util.abort('Failed to convert to WAV file')
         Util.print('OK')
@@ -242,21 +227,35 @@ class App(object):
 
         Util.print_no_lf('Creating output MP3 file')
         merge_cmd = ['ffmpeg', '-y',
-                     '-i', os.path.join(self.tmp_dir, source_wav_file_name),
-                     '-i', os.path.join(self.tmp_dir, 'speech.wav'),
+                     '-i', os.path.join(tmp_dir, source_wav_file_name),
+                     '-i', os.path.join(tmp_dir, 'speech.wav'),
                      '-filter_complex', 'amerge', '-ac', '2', '-c:a', 'libmp3lame',
                      '-q:a', '{}'.format(out_file_info['quality']),
-                     self.args.file_out]
+                     job.file_out]
         if Util.execute_rc(merge_cmd) != 0:
             Util.abort('Failed to create final MP3 file')
         Util.print('OK')
 
         # cleanup
-        if not self.args.no_cleanup:
-            self.cleanup()
+        if not job.args.no_cleanup:
+            if job.args.no_cleanup and tmp_dir is not None:
+                shutil.rmtree(tmp_dir)
         else:
-            if self.args.debug:
-                Util.debug('Leaving "{}" folder intact'.format(self.tmp_dir))
+            if job.args.debug:
+                Util.debug('Leaving "{}" folder intact'.format(tmp_dir))
+
+    def main(self):
+        self.check_env()
+
+        # parse common line arguments
+        self.args = Args.parse_args()
+        Util.init(self.args)
+
+        for file_in in self.args.files_in:
+            if not os.path.isfile(file_in):
+                Util.abort('File not found: {}'.format(file_in))
+
+            self.process_file(Job(self.args, file_in))
 
 
 # ---------------------------------------------------------------------------------