+++ CHANGES.md
@@ -1,3 +1,7 @@
+@dev
+----
+* Added support for batch processing of multiple files in one run
+
 v1.0.0 (2018-10-01)
 -------------------
 * Initial release